#50 localhook

+++ autodoc.sh
@@ -4,22 +4,35 @@
 # Expects esdoc and git to be setup.
 
 if [ "$(id -u)" == "0" ]; then
-   echo "This script should not be run as root. " 1>&2
+   echo "This script should not be run as root!" 1>&2
    exit 1
 fi
 
-TOP=$(cd $(dirname $0) && pwd -L)
+if [[ -n $(git status --porcelain) ]]
+then
+	echo "Repository should be clean!"
+	exit 1
+fi
+
+DATE=`date +%Y-%m-%d:%H:%M:%S`
+BRANCH=$(git rev-parse --abbrev-ref HEAD)
 
 # Generate documentation.
-git checkout gh-pages
-git merge development_core --no-commit --no-ff
-git reset HEAD
-esdoc
-ls -A | grep -v '^doc$\|^document.sh$\|.git$' | xargs rm -r
-mv -v ./doc/dist/* .
-rm doc -r
-rm document.sh
-git add --all
-DATE=`date +%Y-%m-%d:%H:%M:%S`
-git commit -m "AUTODOC: Documentation updated on $DATE"
-git push
+if [[ "$BRANCH" = "master" ]]
+then
+	echo "On master => Pushing documentation to gh-pages"
+	git checkout gh-pages
+	git merge $BRANCH --no-commit --no-ff
+	git reset HEAD
+	esdoc
+	ls -A | grep -v '^doc$\|.git$' | xargs rm -r
+	mv -v ./doc/dist/* .
+	rm doc -r
+	git add --all
+	git commit -m "AUTODOC: Documentation updated on $DATE"
+	git push
+	git checkout $BRANCH
+else
+	echo "Not master => Building local documentation of $BRANCH."
+	esdoc
+fi
